requirements.txt
python-dotenv
httpx
pydantic
requests
beautifulsoup4
lxml
fake-useragent
pyyaml
opencv-python
pandas
Pillow
markdown2
duckduckgo_search
aiohttp
plotly
moviepy
streamlit
python-pptx
pdf2image
python-pptx-interface
